# protocol/patricia: remove error case in Delete