@kr 2017-02-17
Sorry about the formatting of the benchmark comparison. It looks okay in the actual commit message.